+++ R/pirateplot_function.R
@@ -112,7 +112,7 @@ pirateplot <- function(
   avg.line.fun = mean,
   pal = "basel",
   back.col = gray(1),
-  point.cex = 1,
+  point.cex = NULL,
   point.pch = 16,
   point.lwd = 1,
   jitter.val = .03,
@@ -371,6 +371,7 @@ if(theme == 0) {
   if(is.null(avg.line.o)) {avg.line.o <- 0}
   if(is.null(bar.f.o)) {bar.f.o <- 0}
   if(is.null(bar.b.o)) {bar.b.o <- 0}
+  if(is.null(point.cex)) {point.cex <- 1}
 
 }
 
@@ -384,12 +385,12 @@ if(theme == 1) {
   if(is.null(avg.line.o)) {avg.line.o <- 1}
   if(is.null(bar.f.o)) {bar.f.o <- .1}
   if(is.null(bar.b.o)) {bar.b.o <- 0}
-
+  if(is.null(point.cex)) {point.cex <- 1}
 }
 
 if(theme == 2) {
 
-if(is.null(point.o)) {point.o <- 0}
+if(is.null(point.o)) {point.o <- .3}
 if(is.null(bean.b.o)) {bean.b.o <- 0}
 if(is.null(bean.f.o)) {bean.f.o <- .5}
 if(is.null(inf.f.o))  {inf.f.o <- .5}
@@ -401,7 +402,8 @@ if(is.null(bar.b.o))  {bar.b.o <- 0}
 if(is.null(inf.f.col)) {inf.f.col <- "white"}
 if(is.null(inf.b.col)) {inf.b.col <-  "black"}
 if(is.null(avg.line.col)) {avg.line.col <- "black"}
-
+if(is.null(point.col)) {point.col <- "black"}
+if(is.null(point.cex)) {point.cex <- .5}
 }
 }